requirements.txt
alembic==1.6.5
bcrypt==3.2.0
blinker==1.4
cffi==1.15.0
click==8.0.3
colorama==0.4.4
dnspython==2.1.0
dominate==2.6.0
email-validator==1.1.3
Flask==2.0.2
Flask-Bcrypt==0.7.1
Flask-Bootstrap==3.3.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
greenlet==1.1.2
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
Mako==1.1.5
MarkupSafe==2.0.1
mimesis==5.1.1
Pillow==8.3.2
pycparser==2.21
python-dateutil==2.8.2
python-dotenv==0.19.2
python-editor==1.0.4
six==1.16.0
SQLAlchemy==1.4.28
visitor==0.1.3
Werkzeug==2.0.2
WTForms==3.0.0
